Rename the "Operation" class to "Operator"
The term operation in the Coordinate Operations package of the current
ISO-19111 and OGC Topic 2 standards, entangles two related concepts:

- Operator, i.e. an entity able to operate on an operand
- Operation, i.e. the act of applying an operator to an operand

Here, we attempt to untangle the two, by explicitly introducing the
class Operator for objects able to operate on an operand.
The actual operation is carried out using the `Context`-method `apply()`,
which applies the operator to an operand (of base class CoordinateSet).

Hence, with this commit in place, the collection of disentangled
concepts is as follows:

- OperatorMethod: The mathematical expression(s) behind an operator
- Operator: An instantiation of an OperatorMethod and its associated
  defining constants
- Operand: The CoordinateSet operated on by an Operator
- Operation: The act of applying the Operator to the operand